Updates to Solidity 0.6.4
I did this for my own needs, figured I would offer the changeset back upstream in case it is desired.  Feel free to close if you don't want this.  The indentation in this file is a bit messed up, I fixed it in my local copy but restored it to its bad state here so the diff comes out clean.